added a description
To run multiple tasks that doesn’t depend on each other and when they all finish do something else you can use async.parallel. The signature is async.parallel(tasks, callback), where tasks is an array of functions. It will immediately run all the functions in parallel, wait for all of them to call their task callback, and finally when all tasks are complete it will run callback (the final callback).
An example could be to load a forum user’s profile with their details and a list of all their posts.
As input we get the user’s ID, so we can easily get both user details and posts independently of each other (in parallel).